requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
captionbot==0.1.5
certifi==2019.6.16
chardet==3.0.4
discord.py==1.2.2
humanfriendly==4.18
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pyreadline==2.1
requests==2.22.0
typing-extensions==3.7.4
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
